% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{beta-stats}
\alias{beta-stats}
\alias{beta_mean}
\alias{beta_var}
\alias{beta_mode}
\alias{beta_summary}
\title{Summary statistics of a beta distribution}
\usage{
beta_mean(x, shape1 = NULL, shape2 = NULL)

beta_var(x, shape1 = NULL, shape2 = NULL)

beta_mode(x, shape1 = NULL, shape2 = NULL)

beta_summary(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{shape1, shape2}{non-negative parameters of the Beta distribution.}

\item{...}{optional arguments passed to \code{\link[=fitbeta]{fitbeta()}}}
}
\value{
numeric.
}
\description{
Summary statistics of a beta distribution
}
\examples{
set.seed(20250411)
x <- stats::rbeta(1e5, 3.43, 5.55)
beta_summary(x)
}
\references{
https://en.wikipedia.org/wiki/Beta_distribution
}
\seealso{
\code{\link[=fitbeta]{fitbeta()}}, \link[stats:Beta]{stats::Beta}
}
