% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_paths.R
\name{clean_HeFTy}
\alias{clean_HeFTy}
\title{Remove orphan paths}
\usage{
clean_HeFTy(x, min_time = 0)
}
\arguments{
\item{x}{data.frame with time, temperature and segment columns}

\item{min_time}{numeric. Minimum time to consider a path valid (default is 0).
Paths with segments that only exist at times greater than \code{min_time} will be removed.}
}
\value{
data.frame with orphan paths removed.
}
\description{
Sometimes HeFTy produces paths that do not start at the time = 0.
These "orphan" paths are not valid and should be removed. This function removes
paths that only exist at times greater than \code{min_time} (default = 0).
}
