% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{cluster_paths}
\alias{cluster_paths}
\title{Cluster thermal histories}
\usage{
cluster_paths(
  x,
  cluster,
  dist = c("Hausdorff", "Frechet", "Euclidean"),
  method = c("hclust", "kmeans", "pam"),
  ...
)
}
\arguments{
\item{x}{t-T and GOF data of the modeled paths. Output of \code{\link[=read_hefty]{read_hefty()}}.}

\item{cluster}{an integer scalar or vector with the desired number of groups}

\item{dist}{character; algorithm to calculate a dissimilarity matrix
(distance) for lines; one of \code{Euclidean}, \code{Hausdorff} (the default) or
\code{Frechet}.}

\item{method}{character. Clustering method to use; one of \code{"hclust"}
(for \code{\link[stats:hclust]{stats::hclust()}}, the default), \code{"kmeans"} (for \code{\link[stats:kmeans]{stats::kmeans()}}),
or \code{"pam"} (for \code{\link[cluster:pam]{cluster::pam()}}.}

\item{...}{additional arguments passed to cluster method.}
}
\value{
a tibble with the path \code{segment} (integer) and \code{cluster} (factor)
}
\description{
Groups t-T paths into "path families" based on the \emph{Hausdorff distance}
between paths.
}
\examples{
\dontrun{
data(tT_paths1)
tT_paths_subset <- subset(tT_paths1, Comp_GOF >= 0.5)
cluster_paths(tT_paths_subset, cluster = 3)
}
}
