% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{cluster_tendency}
\alias{cluster_tendency}
\title{Cluster tendency of thermochronology cooling paths}
\usage{
cluster_tendency(x, m = NULL, method = c("simple", "torus"))
}
\arguments{
\item{x}{either an object of class \code{"tTdiss"} (output of \code{\link[=path_diss]{path_diss()}}) or a
distance matrix.}

\item{m}{integer. Number of rows to sample from \code{x}. If \code{NULL}, all
rows are considered.}

\item{method}{character. Either \code{"simple"} (default) or \code{"torus"}.}
}
\value{
numeric. The value of Hopkins statistic in the range of 0 and 1 and
its p-value (under the null hypothesis of spatial randomness).
}
\description{
Calculate the Hopkins statistic of Hausdorff or Fréchet distance matrices
to check clusterability of thermochronologic cooling paths.
If the value is close to 1 (far above 0.5), then the
dataset is significantly clusterable.
}
\details{
The algorithm calculates the Hopkins statistic on a transformed
Hausdorff or Fréchet distance matrix using metric multidimensional scaling.
}
\examples{
# example data
data(tT_paths1)
tT_paths_subset <- subset(tT_paths1$paths, Comp_GOF >= 0.4)

# calculate the dissimilarities of the paths:
tT_diss <- path_diss(tT_paths_subset, densify = 1)

set.seed(20250411)
cluster_tendency(tT_diss$diss) # H=0.867, p=0.003
}
