% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_paths.R
\name{crop_paths}
\alias{crop_paths}
\title{Crop cooling paths by time and temperature}
\usage{
crop_paths(x, time = c(-Inf, Inf), temperature = c(-Inf, Inf))
}
\arguments{
\item{x}{an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}})}

\item{time}{numeric. two column vector of the desired time range.}

\item{temperature}{numeric. two column vector of the desired temperature range.}
}
\value{
object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}})
}
\description{
Crop cooling paths by time and temperature
}
\note{
Only the \code{path} slot of \code{x} will be modified.
}
\examples{
# example data
data(tT_paths1)
crop_paths(tT_paths1, time = c(-Inf, 300), temperature = c(-Inf, 200))
}
