% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify_paths.R
\name{densify_paths}
\alias{densify_paths}
\title{Densify Paths}
\usage{
densify_paths(
  x,
  GOF_rank = 10L,
  n = 10L,
  max_distance = 1,
  samples = 100L,
  replace = FALSE
)
}
\arguments{
\item{x}{t-T and GOF data of the modeled paths. Output of \code{\link[=read_hefty]{read_hefty()}}.}

\item{GOF_rank}{numeric. Selects only the \code{GOF_rank}-th highest GOF ranked paths.
If all GOFs should be used, set to \code{Inf}. Default is \code{10}.}

\item{n}{integer. Adds \code{n} (10 by
default) equally-spaced extra points along each path
segment (between vertices).}

\item{max_distance}{numeric. Adds points at a maximum distance of
\code{max_distance} (in Myr) from each other. 1 by default.}

\item{samples}{integer or character. Number of random samples of the data.
This number should be less or equal then the amount of paths.
The default is \code{100}. Paths will be randomly selected \strong{after} the data has
been filtered by the \code{GOF_rank()}.
Optional, set \code{samples} to \code{'all'} to consider all paths ignoring the
\code{GOF_rank()} filter (this sets \code{GOF} to \code{Inf}).
Set \code{samples} to \code{GOF} to consider all paths after the \code{GOF_rank} filter.}

\item{replace}{logical. Should sampling be with replacement?}
}
\value{
tibble
}
\description{
Adds extra points between the vertices of the path segments
}
\note{
A large sample number \code{n} will require a \strong{long(!)}
processing time for this function and subsequent methods such as
\code{\link[=plot_path_density]{plot_path_density()}} or \code{\link[=cluster_paths]{cluster_paths()}}.

If only paths within a specified GOF range should be densified, create a
subset of the data beforehand using either \code{\link[=subset]{subset()}} or \code{\link[dplyr:filter]{dplyr::filter()}}.
}
\examples{
data(tT_paths1)
densify_paths(tT_paths1)
}
