% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fitbeta}
\alias{fitbeta}
\title{Fit of beta distribution to data}
\usage{
fitbeta(x, range = c(0, 1), na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector.}

\item{range}{numeric. two-element vector specifying the interval of the beta distribution}

\item{na.rm}{a logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
mean, mode, and median (sorted by their values), and the variance
of the beta distribution fitted to \code{x}.
}
\description{
Estimates the shape parameters for a beta distribution using
}
\examples{
set.seed(20250411)
x <- stats::rbeta(1e5, 3.43, 5.55)
fitbeta(x)
}
\references{
https://en.wikipedia.org/wiki/Beta_distribution
}
\seealso{
\link{beta-stats}, \link[stats:Beta]{stats::Beta}
}
