% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path_density.R
\name{gof_weighting}
\alias{gof_weighting}
\title{Create weightings from goodness-of-fit values}
\usage{
gof_weighting(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector of GOF values (between 0 and 1)}

\item{na.rm}{logical. Whether missing values should be removed (\code{FALSE} by default.)}
}
\value{
numeric vector of weights between 0 and 1
}
\description{
HeFTy goodness-of-fit (GOF) values range from 0 (worst fit) to 1 (best fit).
Because, GOF values greater than 0.5 are considered equally good ("acceptable"),
this function rescales the GOF values to weighting values.
}
\examples{
gof_values <- c(0.1, 0.3, 0.5, 0.7, 0.9)
gof_weighting(gof_values)
}
\seealso{
\code{\link[=plot_path_density_filled_weighted]{plot_path_density_filled_weighted()}} for calculating weighted density of t-T paths.
}
