% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_distances.R
\name{hausdorff}
\alias{hausdorff}
\alias{directed_hausdorff}
\title{Hausdorff Distance}
\usage{
hausdorff(P, Q)

directed_hausdorff(P, Q)
}
\arguments{
\item{P, Q}{numerical matrices, representing points in an m-dim. space.}
}
\value{
A single scalar, the Hausdorff distance (dimension).
}
\description{
Hausdorff distance (aka Hausdorff dimension)
}
\examples{
P <- matrix(c(1, 1, 2, 2, 5, 4, 5, 4), 4, 2)
Q <- matrix(c(4, 4, 5, 5, 2, 1, 2, 1), 4, 2)
hausdorff(P, Q)
}
