% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{use_parallel}
\alias{use_parallel}
\title{Helper function to set up parallel workflows depending on the OS}
\usage{
use_parallel(use = TRUE, workers = NULL)
}
\arguments{
\item{use}{logical. Enable (\code{TRUE}) or disable (\code{FALSE}) parallel computation.}

\item{workers}{The number of parallel processes to use.}
}
\description{
Helper function to set up parallel workflows depending on the OS
}
