% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{path_diss}
\alias{path_diss}
\title{Dissimilarity of thermochronology cooling paths}
\usage{
path_diss(x, dist = c("Hausdorff", "Frechet"), densify = 0, simplify = 0, ...)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}) or
a \code{data.frame} containing the \code{time}, \code{temperature} columns of the modeled paths.}

\item{dist}{character. Algorithm to calculate a dissimilarity matrix
(distance) for lines; either \code{Hausdorff} (the default) or \code{Frechet}.}

\item{densify}{numeric. optionally use a value
between 0 and 1 to densify the geometry description. Default is 0.}

\item{simplify}{numeric. Optional tolerance parameter to specify the
amount of simplification of path geometries in order to save processing time.
The tolerance is in the same unit as temperature and should relate to the
expected temperature spread.
The simplification is applied before the densification.
Some good values are 0.5, 2.5, 5, and 10. Default is 0.}

\item{...}{(optional) arguments passed to \code{\link[=cluster_tendency]{cluster_tendency()}}.}
}
\value{
\code{tTdiss} object, i.e. a list containing
\describe{
\item{\code{paths}}{t-T paths as \code{sf} object in time-temperature space}
\item{\code{diss}}{the \eqn{n \times n} dissimilarity matrix (\eqn{n} is number of paths)}
\item{\code{method}}{the dissimilarity algorithm used}
\item{\code{hopkins}}{the Hopkins statistic and its p-value.}
}
}
\description{
Calculates the dissimilarity matrix and Hopkins statistic for
thermochronology cooling paths using the \emph{Hausdorff} or the \emph{Fréchet distance}.
}
\details{
The Hausdorff distance is the greatest of all the distances from a
point in one set to the closest point in the other set.
The Fréchet distance additionally takes into account the location and
ordering of the points along the curves (the "flow").
}
\examples{
# example data
data(tT_paths1)
tT_paths_subset <- subset(tT_paths1$paths, Comp_GOF >= 0.2)

# calculate the dissimilarities of the paths:
tT_diss <- path_diss(tT_paths_subset, densify = 1)
print(tT_diss)

# the `diss` object of the output can be used for clustering, e.g.:
stats::kmeans(tT_diss$diss, centers = 3)
}
