% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{path_hcut}
\alias{path_hcut}
\title{Hierarchical Clustering of Cooling Paths}
\usage{
path_hcut(x, k, FUN = stats::hclust, ...)
}
\arguments{
\item{x}{an object of class \code{"HeFTy"}, \verb{"tTdiss}" or \code{"dist"} (dissimilarity matrix).}

\item{k}{integer. number of clusters to be generated}

\item{FUN}{hierarchical clustering function to be used, i.e. one of
\code{\link[stats:hclust]{stats::hclust()}} (the default), \code{\link[cluster:agnes]{cluster::agnes()}}, \code{\link[cluster:diana]{cluster::diana()}}).}

\item{...}{optional arguments past to \code{hc_func}}
}
\value{
an object of class \code{"hcut"} containing the result of the standard
function used (read the documentation of \code{\link[stats:hclust]{stats::hclust()}}, \code{\link[cluster:agnes]{cluster::agnes()}}, \code{\link[cluster:diana]{cluster::diana()}}).

It includes also
\describe{
\item{cluster}{the cluster assignment of observations after cutting the tree}
\item{nbclust}{the number of clusters}
\item{size}{the size of clusters}
}
}
\description{
Convenience function to compute hierarchical clustering and cut the tree into k clusters
}
\examples{
data(tT_paths1)
tT_paths_subset <- subset(tT_paths1$paths, Comp_GOF >= 0.4)

# calculate the dissimilarities of the paths:
tT_diss <- path_diss(tT_paths_subset)
path_hcut(tT_diss$diss, 3)
}
