% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_cluster.R
\name{path_nbclust}
\alias{path_nbclust}
\title{Optimal number of clusters}
\usage{
path_nbclust(
  x,
  FUNcluster = path_hcut,
  k.max = 10,
  n.threshold = Inf,
  linecolor = "#B63679FF",
  ...
)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}),
\verb{"tTdiss} (output of \code{\link[=path_diss]{path_diss()}}), or
a \code{data.frame} containing the \code{time}, \code{temperature}, and \code{segment} columns of the modeled paths.}

\item{FUNcluster}{cluster function. Default is \code{\link[=path_hcut]{path_hcut()}}.}

\item{k.max}{integer. the maximum number of clusters to consider, must be at least two.}

\item{n.threshold}{integer. If the number of paths is greater than this value,
a random sample of size \code{n.threshold} of paths will be used to determine the optimal number of clusters.
Must be greater than \code{k.max}. Default is \code{Inf}, i.e., no sampling.
Consider to specify this parameter if processing takes too much time.}

\item{linecolor}{line color}

\item{...}{optionally further arguments for \code{FUNcluster}}
}
\value{
list containing the optimal number
of clusters (\code{optimal}), and a plot showing the average silhouette width for each number of clusters (\code{plot}).
}
\description{
Determines the optimal number of clusters in a
dissimilarity matrix using \emph{average silhouette width}.
}
\examples{
# example data
data(tT_paths1)
res <- path_nbclust(tT_paths1, n.threshold = 500)
res$optimal

tT_paths_subset <- subset(tT_paths1$paths, Comp_GOF >= 0.4)
res <- path_nbclust(tT_paths_subset)
res$optimal
}
