% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paths.R
\name{plot_paths}
\alias{plot_paths}
\title{Plot the thermochronology paths}
\usage{
plot_paths(x, cluster = NULL, pal = viridisLite::viridis, breaks = 5, ...)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}) or
a \code{data.frame} containing the \code{time}, \code{temperature}, and \code{segment} columns
of the modeled paths.}

\item{cluster}{(optional) data.frame containing the \code{segment} and \code{cluster}
columns to be merged with \code{x}.}

\item{pal}{color function}

\item{breaks}{breaks to cut the GOF values}

\item{...}{options passed to \code{pal}}
}
\value{
plot
}
\description{
Plots the time-temperature paths. If the input \code{x} is a \code{"HeFTy"} object or a
data.frame containing the goodness of fit values (as. \code{Comp_GOF} column),
the paths are color coded by the goodness of fit. If \code{cluster} is specified,
the paths are color coded by the clusters.
}
\examples{
# example data
data(tT_paths1)
tT_paths1$paths <- subset(tT_paths1$paths, Comp_GOF >= 0.4)

# Plot the paths
plot_paths(tT_paths1)

# Show path clusters
plot_paths(tT_paths1, cluster = cluster_paths(tT_paths1, k = 3))
}
