% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path_density.R
\name{plt_density}
\alias{plt_density}
\alias{plot_path_density_filled}
\alias{plot_path_density}
\alias{plot_path_density_filled_weighted}
\title{Path Density Plot}
\usage{
plot_path_density_filled(
  x,
  bins = 50L,
  densify = TRUE,
  show.legend = NA,
  geom = "density_2d_filled",
  n = 100L,
  ...
)

plot_path_density(
  x,
  bins = 50L,
  densify = TRUE,
  show.legend = NA,
  n = 100L,
  ...
)

plot_path_density_filled_weighted(
  x,
  bins = 50L,
  densify = TRUE,
  show.legend = NA,
  n = 100L,
  weights = NULL,
  ...
)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}),
a \code{data.frame} containing the \code{time}, \code{temperature} columns of the modeled
paths, or output of \code{\link[=densify_paths]{densify_paths()}}.}

\item{bins}{integer. Amount of bins used for the kernel density estimate. 50 by default.}

\item{densify}{logical. Whether the paths in \code{x} should be densified first?
Default is \code{TRUE}.}

\item{show.legend}{logical. Should this layer be included in the legends? \code{NA},
the default, includes if any aesthetics are mapped. \code{FALSE} never includes,
and \code{TRUE} always includes. It can also be a named logical vector to finely
select the aesthetics to display.}

\item{geom}{Use to override the default connection between
\code{\link[ggplot2:geom_density_2d]{ggplot2::geom_density_2d()}} and \code{\link[ggplot2:geom_density_2d]{ggplot2::stat_density_2d()}}.
For more information at overriding these connections, see how the stat and geom arguments work.}

\item{n}{integer. Number of grid points in each direction.}

\item{...}{Arguments passed on to \code{\link[=densify_paths]{densify_paths()}} (only if \code{densify=TRUE}).}

\item{weights}{numeric vector. Weights for each path segment.}
}
\value{
ggplot
}
\description{
Creates a 2d kernel density estimate of the t-T paths and plots it using
ggplot
}
\examples{
data(tT_paths1)
plot_path_density(tT_paths1)
plot_path_density_filled(tT_paths1)
plot_path_density_filled(tT_paths1, geom = "raster")
plot_path_density_filled_weighted(tT_paths1, weights = gof_weighting(tT_paths1$paths$Comp_GOF))
}
\seealso{
\code{\link[=gof_weighting]{gof_weighting()}} for rescaling goodness-of-fit values to weights.
}
