% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hefty.R
\name{read_hefty}
\alias{read_hefty}
\title{Import inverse thermal history model}
\usage{
read_hefty(fname)
}
\arguments{
\item{fname}{path to the .txt file that contains the HeFTy outputs}
}
\value{
\code{list} of the individual paths, the constraints, the weighted mean
path and the grain summary statistics.
}
\description{
Read time, temperature and GOF data of modeled thermal history paths out of
HeFTy output .txt file
}
\examples{
path2myfile <- system.file("112-9-30-zr-inv.txt", package = "thermochron")
read_hefty(path2myfile)
}
