% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hefty.R
\name{read_hefty_xlsx}
\alias{read_hefty_xlsx}
\title{Read time, temperature and GOF data from HeFTy output from excel file}
\usage{
read_hefty_xlsx(fname)
}
\arguments{
\item{fname}{path to the excel spreadsheet that contains the HeFTy outputs, i.e. the t-T-paths in sheet 1 and the GOF values in sheet 2}
}
\value{
\code{data.frame} of the combined data
}
\description{
Read time, temperature and GOF data from HeFTy output from excel file
}
\examples{
\dontrun{
path2myfile <- system.file("s14MM_v1.xlsx", package = "HeFTy.SmoothR")
read_hefty_xlsx(path2myfile)
}
}
\seealso{
\code{\link[=read_hefty]{read_hefty()}}
}
