% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setup_parallel}
\alias{setup_parallel}
\title{Helper function to set up parallel workflows depending on the OS}
\usage{
setup_parallel(workers = NULL)
}
\arguments{
\item{workers}{The number of parallel processes to use.}
}
\description{
Helper function to set up parallel workflows depending on the OS
}
