% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoclustr.R
\docType{package}
\name{thermoclustr-package}
\alias{thermoclustr}
\alias{thermoclustr-package}
\title{thermoclustr: Analyzing Inverse Thermal History Models}
\description{
Density estimates and cluster methods for HeFTy-modeled time-temperature cooling paths.
}
\author{
\strong{Maintainer}: Tobias Stephan \email{tobias.stephan1@yahoo.com} (\href{https://orcid.org/0000-0002-9290-014X}{ORCID})

Authors:
\itemize{
  \item Taís Fontes Pinto \email{tais.fontespinto@ucalgary.ca} [copyright holder]
}

}
\keyword{internal}
