% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tT_paths}
\alias{tT_paths}
\title{HeFTy paths example}
\format{
An object of class \code{tibble}
}
\usage{
data('tT_paths')
}
\description{
Modelled time-temperature paths from the HeFTy model.
}
\examples{
data("tT_paths")
head(tT_paths)
}
\keyword{datasets}
