% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tT_paths1}
\alias{tT_paths1}
\title{HeFTy paths example 1}
\format{
An object of class \code{tibble}
}
\usage{
data('tT_paths1')
}
\description{
Modelled time-temperature paths from the HeFTy model.
}
\examples{
data("tT_paths1")
head(tT_paths1)
}
\keyword{datasets}
