% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tT_paths2}
\alias{tT_paths2}
\title{HeFTy paths example 2}
\format{
An object of class \code{tibble}
}
\usage{
data('tT_paths2')
}
\description{
Modelled time-temperature paths from the HeFTy model.
}
\examples{
data("tT_paths2")
head(tT_paths2)
}
\keyword{datasets}
