% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{cluster_paths}
\alias{cluster_paths}
\title{Cluster thermal histories}
\usage{
cluster_paths(
  x,
  cluster,
  dist = c("Hausdorff", "Frechet"),
  method = c("hclust", "kmeans", "pam", "specc", "dbscan", "hdbscan", "diana", "agnes",
    "mona", "clara", "fanny"),
  ...
)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}), an
object of \code{"tTdiss"} (output of \code{\link[=path_diss]{path_diss()}}),
or a \code{data.frame} containing the \code{time}, \code{temperature} columns of the
modeled paths.}

\item{cluster}{an integer scalar or vector with the desired number of groups.
Ignored when \code{dist} equal to \code{dbscan} or \code{hdbscan} (\strong{see Note}).}

\item{dist}{character. Algorithm to calculate a dissimilarity matrix
(distance) for lines; either \code{Hausdorff} (the default), or
\code{Frechet}.}

\item{method}{character. Clustering method to use. Currently implemented are
\describe{
\item{\code{"hclust"}}{for Hierarchical Clustering using \code{\link[stats:hclust]{stats::hclust()}}, the default)}
\item{\code{"kmeans"}}{for K-Means Clustering using \code{\link[stats:kmeans]{stats::kmeans()}})}
\item{\code{"pam"}}{for Partitioning Around Medoids using \code{\link[cluster:pam]{cluster::pam()}}}
\item{\code{"dbscan"}}{for Density-based Spatial Clustering of Applications with
Noise (DBSCAN) using \code{\link[dbscan:dbscan]{dbscan::dbscan()}} (\strong{see Note})}
\item{\code{"hdbscan"}}{for Hierarchical DBSCAN using \code{\link[dbscan:hdbscan]{dbscan::hdbscan()}}
(\strong{see Note})}
\item{\code{"specc"}}{for Spectral Clustering using \code{\link[kernlab:specc]{kernlab::specc()}}}
\item{\code{"agnes"}}{for agglomerative hierarchical clustering using \code{\link[cluster:agnes]{cluster::agnes()}}}
\item{\code{"diana"}}{for divisive hierarchical clustering using \code{\link[cluster:diana]{cluster::diana()}}}
\item{\code{"clara"}}{Clustering Large Applications using \code{\link[cluster:clara]{cluster::clara()}}}
\item{\code{"fanny"}}{Fuzzy Analysis Clustering using \code{\link[cluster:fanny]{cluster::fanny()}}}
}}

\item{...}{additional arguments passed to cluster method.}
}
\value{
a tibble with the path \code{segment} (integer) and \code{cluster} (factor)
}
\description{
Groups t-T paths into "path families" based on the \emph{Hausdorff} or
\emph{Fréchet distance} between paths.
}
\note{
that \code{dbscan} and \code{hdbscan} methods require \code{eps} and \code{minPts} arguments.
Optimal \code{eps} values can be visually estimated from the "knee" in a k-nearest
neighbor distance plot using:
\code{\link[dbscan:kNNdist]{dbscan::kNNdistplot()}}
}
\examples{
\dontrun{
# example data
data(tT_paths1)
tT_paths_subset <- subset(tT_paths1$paths, Comp_GOF >= 0.5)

# cluster the paths
cluster_paths(tT_paths_subset, cluster = 3)
}
}
\seealso{
\code{\link[=path_diss]{path_diss()}}
}
