% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{cluster_paths}
\alias{cluster_paths}
\title{Clustering thermal histories models}
\usage{
cluster_paths(
  x,
  cluster,
  dist = c("Hausdorff", "Frechet"),
  method = c("hclust", "kmeans", "pam", "specc", "dbscan", "hdbscan", "diana", "agnes",
    "clara", "fanny"),
  naming = c("asis", "GOF", "size"),
  ...
)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}), an
object of \code{"tTdiss"} (output of \code{\link[=path_diss]{path_diss()}}),
or a \code{data.frame} containing the \code{time}, \code{temperature}, and \code{segment} columns of the
modeled paths.}

\item{cluster}{an integer scalar or vector with the desired number of groups.
Ignored when \code{dist} equal to \code{dbscan} or \code{hdbscan} (\strong{see Note}).}

\item{dist}{character. Algorithm to calculate a dissimilarity matrix
(distance) for lines; either \code{Hausdorff} (the default), or
\code{Frechet}.}

\item{method}{character. Clustering method to be applied. Currently implemented are
\describe{
\item{\code{"hclust"}}{Hierarchical Clustering using \code{\link[stats:hclust]{stats::hclust()}}, the default)}
\item{\code{"kmeans"}}{K-Means Clustering using \code{\link[stats:kmeans]{stats::kmeans()}})}
\item{\code{"pam"}}{Partitioning Around Medoids using \code{\link[cluster:pam]{cluster::pam()}}}
\item{\code{"dbscan"}}{Density-based Spatial Clustering of Applications with
Noise (DBSCAN) using \code{\link[dbscan:dbscan]{dbscan::dbscan()}} (\strong{see Note})}
\item{\code{"hdbscan"}}{Hierarchical DBSCAN using \code{\link[dbscan:hdbscan]{dbscan::hdbscan()}}
(\strong{see Note})}
\item{\code{"specc"}}{Spectral Clustering using \code{\link[kernlab:specc]{kernlab::specc()}}}
\item{\code{"agnes"}}{Agglomerative hierarchical clustering using \code{\link[cluster:agnes]{cluster::agnes()}}}
\item{\code{"diana"}}{Divisive hierarchical clustering using \code{\link[cluster:diana]{cluster::diana()}}}
\item{\code{"clara"}}{Clustering Large Applications using \code{\link[cluster:clara]{cluster::clara()}}}
\item{\code{"fanny"}}{Fuzzy Analysis Clustering using \code{\link[cluster:fanny]{cluster::fanny()}}}
}}

\item{naming}{character. Naming scheme for clusters. One of
\code{"asis"} (output of underlying cluster algorithm),
\code{"GOF"} (ranks of the mean GOF values within clusters), and
\code{"size"} (ranks of the size of clusters).
Outliers detected by \code{dbscan} of \code{hdbscan} will always named as \code{0}.}

\item{...}{additional arguments passed to cluster method.}
}
\value{
a tibble with the path \code{segment} (integer) and \code{cluster} (factor)
}
\description{
Groups t-T paths into "path families" based on the \emph{Hausdorff} or
\emph{Fréchet distance} between paths.
}
\details{
If you want to use a different clustering method that is not built
in the current version of \code{thermochron}, you can use the dist ance matrix
produced by \code{\link[=path_diss]{path_diss()}} and feed your cluster algorithm.
}
\note{
that \code{dbscan} and \code{hdbscan} methods require \code{eps} and \code{minPts} arguments.
Optimal \code{eps} values can be visually estimated from the "knee" in a k-nearest
neighbor distance plot using \code{\link[dbscan:kNNdist]{dbscan::kNNdistplot()}}'.
}
\examples{
# example data
data(tT_paths1)
tT_paths1$paths <- subset(tT_paths1$paths, Comp_GOF >= 0.4)

# cluster the paths
cluster_paths(tT_paths1, cluster = 3)
}
\seealso{
\code{\link[=path_diss]{path_diss()}} for calculating dissimilarities and \code{\link[=path_nbclust]{path_nbclust()}}
for determing the optimal number of clusters.
}
