% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{cluster_paths}
\alias{cluster_paths}
\title{Clustering thermal histories models}
\usage{
cluster_paths(
  x,
  k,
  dist = c("Hausdorff", "Frechet"),
  method = c("hclust", "kmeans", "pam", "dbscan", "hdbscan", "specc", "diana", "agnes",
    "clara", "fanny"),
  naming = c("asis", "GOF", "size"),
  warn = TRUE,
  threshold = 0.01,
  ...
)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}), an
object of \code{"tTdiss"} (output of \code{\link[=path_diss]{path_diss()}}),
or a \code{data.frame} containing the \code{time}, \code{temperature}, and \code{segment} columns of the
modeled paths.}

\item{k}{an integer scalar or vector with the desired number of groups.
Ignored when \code{dist} equal to \code{dbscan} or \code{hdbscan} (\strong{see Note}).}

\item{dist}{character. Algorithm to calculate a dissimilarity matrix
(distance) for lines; either \code{Hausdorff} (the default), or
\code{Frechet}.}

\item{method}{character. Clustering method to be applied. Currently implemented are
\describe{
\item{\code{"hclust"}}{Hierarchical Clustering using \code{\link[stats:hclust]{stats::hclust()}}, the default)}
\item{\code{"kmeans"}}{K-Means Clustering using \code{\link[stats:kmeans]{stats::kmeans()}})}
\item{\code{"pam"}}{Partitioning Around Medoids using \code{\link[cluster:pam]{cluster::pam()}}}
\item{\code{"dbscan"}}{Density-based Spatial Clustering of Applications with
Noise (DBSCAN) using \code{\link[dbscan:dbscan]{dbscan::dbscan()}} (\strong{see Note})}
\item{\code{"hdbscan"}}{Hierarchical DBSCAN using \code{\link[dbscan:hdbscan]{dbscan::hdbscan()}}
(\strong{see Note})}
\item{\code{"specc"}}{Spectral Clustering using \code{\link[kernlab:specc]{kernlab::specc()}}}
\item{\code{"agnes"}}{Agglomerative hierarchical clustering using \code{\link[cluster:agnes]{cluster::agnes()}}}
\item{\code{"diana"}}{Divisive hierarchical clustering using \code{\link[cluster:diana]{cluster::diana()}}}
\item{\code{"clara"}}{Clustering Large Applications using \code{\link[cluster:clara]{cluster::clara()}}}
\item{\code{"fanny"}}{Fuzzy Analysis Clustering using \code{\link[cluster:fanny]{cluster::fanny()}}}
}}

\item{naming}{character. Naming scheme for clusters. One of
\code{"asis"} (output of underlying cluster algorithm),
\code{"GOF"} (ranks of the mean GOF values within clusters), and
\code{"size"} (ranks of the size of clusters).
Outliers detected by \code{dbscan} of \code{hdbscan} will always named as \code{0}.}

\item{warn}{logical. Should there be a warning message if at least one cluster
contains less than (\code{threshold} * 100)\% of the total paths?}

\item{threshold}{numeric. The significance threshold as a fraction of the
total amount of paths (\code{0.01} by default). If one cluster contains less
paths per total paths than this value, a warning message is created. Ignored
if \code{warn} is \code{FALSE}.}

\item{...}{additional arguments passed to cluster method.}
}
\value{
a data.frame with the path \code{segment} (integer) and \code{cluster} (factor)
}
\description{
Groups t-T paths into "path families" based on the \emph{Hausdorff} or
\emph{Fréchet distance} between paths.
}
\details{
If you want to use a different clustering method that is not built
in the current version of \code{thermochron}, you can use the distance matrix
produced by \code{\link[=path_diss]{path_diss()}} and feed your cluster algorithm.
}
\note{
that \code{dbscan} and \code{hdbscan} methods require \code{eps} and \code{minPts} arguments.
Optimal \code{eps} values can be visually estimated from the "knee" in a k-nearest
neighbor distance plot using \code{\link[dbscan:kNNdist]{dbscan::kNNdistplot()}}'.
}
\examples{
data(tT_paths)
tT_paths$paths <- subset(tT_paths$paths, Comp_GOF >= 0.4)

# cluster the paths
cluster_paths(tT_paths, k = 3)
}
\seealso{
\code{\link[=path_diss]{path_diss()}} for calculating dissimilarities and \code{\link[=path_nbclust]{path_nbclust()}}
for determining the optimal number of clusters.
}
