% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopkins.R
\name{cluster_tendency}
\alias{cluster_tendency}
\title{Cluster tendency of thermochronology cooling paths}
\usage{
cluster_tendency(x, m = NULL, method = c("simple", "torus"))
}
\arguments{
\item{x}{either an object of class \code{"tTdiss"} (output of \code{\link[=path_diss]{path_diss()}}), a
distance matrix (object \code{"dist"}), or a \code{data.frame} containing coordinates from a dimension-reducing algorithm.}

\item{m}{integer. Number of rows to sample from \code{x}. If \code{NULL}, all
rows are considered.}

\item{method}{character. Either \code{"simple"} (default) or \code{"torus"}.}
}
\value{
numeric. The value of Hopkins statistic in the range of 0 and 1 and
its p-value (under the null hypothesis of spatial randomness).
}
\description{
Calculate the Hopkins statistic of Hausdorff or Fréchet distance matrices
to check clusterability of thermochronologic cooling paths.
Calculated values 0-0.3 indicate regularly-spaced data. Values around 0.5
indicate random data. Values 0.7-1 indicate clustered data.
}
\details{
Calculates the Hopkins statistic on the transformed
Hausdorff or Fréchet distance matrix using metric multidimensional scaling by default.
}
\examples{
data(tT_paths)
tT_paths_subset <- subset(tT_paths$paths, Comp_GOF >= 0.4)

# calculate the dissimilarities of the paths:
tT_diss <- path_diss(tT_paths_subset, densify = 1)

set.seed(20250411)
cluster_tendency(tT_diss) # H=0.77, p=2.88e-11

# Alternative dimension reducing algorithms can be used:
\dontshow{if (require("MASS")) withAutoprint(\{ # examplesIf}
# Kriskal's non-metric MDS
library(MASS)
tT_nmds <- isoMDS(tT_diss$diss)$points
cluster_tendency(tT_nmds)
\dontshow{\}) # examplesIf}
\dontshow{if (require("uwot")) withAutoprint(\{ # examplesIf}
# Uniform Manifold Approximation and Projection (UMAP)  
library(uwot)
tT_umap <- umap2(tT_diss$diss)
cluster_tendency(tT_umap)
\dontshow{\}) # examplesIf}
}
