% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{count_cluster}
\alias{count_cluster}
\title{Count the number of paths in each cluster}
\usage{
count_cluster(x)
}
\arguments{
\item{x}{data.frame with a \verb{$cluster} column}
}
\value{
named array of integers, the number of paths in each cluster
}
\description{
Helper function to calculate the paths in each cluster and give a warning
message if one cluster contains not enough paths to be considered significant
}
\examples{
data(tT_paths)
tT_paths$paths <- subset(tT_paths$paths, Comp_GOF >= 0.4)

# cluster the paths
res <- cluster_paths(tT_paths, k = 3)
count_cluster(res)
}
