% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{path_diss}
\alias{path_diss}
\title{Dissimilarity of t-T paths}
\usage{
path_diss(x, method = c("Hausdorff", "Frechet"), densify = 0)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}) or
a \code{data.frame} containing the \code{time}, \code{temperature} columns of the modeled paths.}

\item{method}{character. Algorithm to calculate a dissimilarity matrix
(distance) for lines; either \code{Hausdorff} (the default) or \code{Frechet}.}

\item{densify}{numeric. optionally use a value
between 0 and 1 to densify the geometry description.}
}
\value{
\code{tTdiss} object, i.e. a list containing
\describe{
\item{\code{paths}}{t-T paths as \code{sf} object in time-temperature space}
\item{\code{diss}}{the \eqn{n \times n} dissimilarity matrix (\eqn{n} is number of paths)}
}
}
\description{
Calculates the dissimilarity matrix between t-T paths using the \emph{Hausdorff} or
\emph{Fréchet distance}
}
\details{
The Hausdorff distance is the greatest of all the distances from a
point in one set to the closest point in the other set.
The Fréchet distance additionally takes into account the location and
ordering of the points along the curves (the "flow").
}
\examples{
# example data
data(tT_paths1)
tT_paths_subset <- subset(tT_paths1$paths, Comp_GOF >= 0.5)

# calculate the dissimilarities of the paths:
tT_diss <- path_diss(tT_paths_subset, densify = 1)
print(tT_diss)

# the `diss` object of the output can be used for clustering, e.g.:
stats::kmeans(tT_diss$diss, centers = 3)
}
