% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{path_diss}
\alias{path_diss}
\title{Dissimilarity of thermochronology cooling paths}
\usage{
path_diss(x, dist = c("Hausdorff", "Frechet"), densify = 0, simplify = 0, ...)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}) or
a \code{data.frame} containing the \code{time}, \code{temperature}, and \code{segment} columns of the modeled paths.}

\item{dist}{character. Algorithm to calculate a dissimilarity matrix
(distance) for lines; either \code{Hausdorff} (the default) or \code{Frechet}.}

\item{densify}{numeric. optionally use a value
between 0 and 1 to densify the geometry description. Default is 0.}

\item{simplify}{numeric. Optional tolerance parameter to specify the
amount of simplification of path geometries in order to save processing time.
The tolerance is in the same unit as temperature and should relate to the
expected temperature spread.
The simplification is applied before the densification.
Some good values are 0.5, 2.5, 5, and 10. Default is 0 (no simplification).}

\item{...}{(optional) arguments passed to \code{\link[=cluster_tendency]{cluster_tendency()}}.}
}
\value{
\code{tTdiss} object, i.e. a list containing
\describe{
\item{\code{paths}}{t-T paths as \code{sf} object in time-temperature space}
\item{\code{diss}}{the \eqn{n \times n} dissimilarity matrix (\eqn{n} is number of paths)}
\item{\code{method}}{the dissimilarity algorithm used}
\item{\code{hopkins}}{the Hopkins statistic and its p-value.}
}
}
\description{
Calculates the dissimilarity matrix and Hopkins statistic for
thermochronology cooling paths using the \emph{Hausdorff} or the \emph{Fréchet distance}.
}
\details{
The Hausdorff distance is the greatest of all the distances from a
point in one set to the closest point in the other set.
The Fréchet distance additionally takes into account the location and
ordering of the points along the curves (the "flow").
}
\note{
The algorithm calculates the pairwise dissimilarities between the n
paths and creates a n x n matrix.
A large number of paths may end up in very long processing times and can
cause memory issues.
Thus, it is recommended to filter the paths beforehand using either
\code{\link[=crop_paths]{crop_paths()}} or \code{\link[=subset]{subset()}} to, for example,
include only paths with large GOF (see examples),
or create a random subset using, e.g., \code{\link[=sample]{sample()}} (see examples).
Additionally,
the \code{simplify} parameter can be adjusted to reduce the number of vertices
along the paths to speed up the processing time.
}
\examples{
# example data
data(tT_paths1)
tT_paths_subset <- subset(tT_paths1$paths, Comp_GOF >= 0.4)

# calculate the dissimilarities of the paths:
tT_diss <- path_diss(tT_paths_subset, densify = 1)

# the `diss` object of the output can be used for clustering using any 
# available cluster algorithm, e.g.:
stats::kmeans(tT_diss$diss, centers = 3)

# using a random subset of paths (in case of too large data):
set.seed(20250411)
random_segments <- sample(unique(tT_paths1$paths$segment), size = 100) # select 100 random path segments
tT_paths_rnd <- subset(tT_paths1$paths, segment \%in\% random_segments)
tT_diss_rnd <- path_diss(tT_paths_rnd)
}
