% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_cluster.R
\name{path_nbclust}
\alias{path_nbclust}
\title{Optimal number of clusters}
\usage{
path_nbclust(
  x,
  FUNcluster = factoextra::hcut,
  k.max = 5,
  nboot = 50,
  dim = 4,
  barfill = "#1D1147FF",
  barcolor = "#1D1147FF",
  linecolor = "#1D1147FF",
  ...
)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}),
\verb{"tTdiss} (output of \code{\link[=path_diss]{path_diss()}}), or
a \code{data.frame} containing the \code{time}, \code{temperature}, and \code{segment} columns of the modeled paths.}

\item{FUNcluster}{cluster function}

\item{k.max}{integer. the maximum number of clusters to consider, must be at least two.}

\item{nboot}{integer. integer, number of Monte Carlo ("bootstrap") samples.
Used only for determining the number of clusters using gap statistic.}

\item{dim}{integer. Number of dimensions for \code{\link[stats:cmdscale]{stats::cmdscale()}} used for the gap statistic, default is 4.}

\item{barfill, barcolor, linecolor}{fill color and outline color for bars}

\item{...}{optionally further arguments for \code{FUNcluster()}}
}
\value{
list containing the results of the three methods, the optimal number of clusters (median of the results), and the plots for each method.
}
\description{
Convenience function to determine the optimal number of clusters in a dissimilarity matrix using three methods
wss, silhouette, and gap statistic.
}
\examples{
# example data
data(tT_paths1)
tT_paths_subset <- subset(tT_paths1$paths, Comp_GOF >= 0.4)
res <- path_nbclust(tT_paths_subset)
res$optimal
}
