% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_cluster.R
\name{path_nbclust}
\alias{path_nbclust}
\title{Optimal number of clusters}
\usage{
path_nbclust(
  x,
  FUNcluster = factoextra::hcut,
  k.max = 10,
  wss = FALSE,
  silhouette = TRUE,
  gap.statistic = FALSE,
  dim = 4,
  n.threshold = Inf,
  linecolor = "#1D1147FF",
  ...
)
}
\arguments{
\item{x}{either an object of class \code{"HeFTy"} (output of \code{\link[=read_hefty]{read_hefty()}}),
\verb{"tTdiss} (output of \code{\link[=path_diss]{path_diss()}}), or
a \code{data.frame} containing the \code{time}, \code{temperature}, and \code{segment} columns of the modeled paths.}

\item{FUNcluster}{cluster function}

\item{k.max}{integer. the maximum number of clusters to consider, must be at least two.}

\item{wss, silhouette, gap.statistic}{logical. Select if this statistic should be included.}

\item{dim}{integer. Number of dimensions for \code{\link[stats:cmdscale]{stats::cmdscale()}} used for the gap statistic, default is 4.}

\item{n.threshold}{integer. If the number of paths is greater than this value,
a random sample of size \code{n.threshold} of paths will be used to determine the optimal number of clusters.
Must be greater than \code{k.max}. Default is \code{Inf}, i.e., no sampling.
Consider to specify this parameter if processing takes too much time.}

\item{linecolor}{line color}

\item{...}{optionally further arguments for \code{\link[factoextra:fviz_nbclust]{factoextra::fviz_nbclust()}}}
}
\value{
list containing the results of the three methods, the optimal number
of clusters (median of the results), and the plots for each method.
}
\description{
Convenience function to determine the optimal number of clusters in a
dissimilarity matrix using the three methods
\emph{within cluster sums of squares}, \emph{average silhouette}, and \emph{gap statistic}.
}
\examples{
# example data
data(tT_paths1)
tT_paths_subset <- subset(tT_paths1$paths, Comp_GOF >= 0.4)
res <- path_nbclust(tT_paths_subset)
res$optimal
}
