% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hefty.R
\name{read_hefty}
\alias{read_hefty}
\title{Import inverse thermal history model}
\usage{
read_hefty(fname, remove_orphans = TRUE)
}
\arguments{
\item{fname}{path to the .txt file that contains the HeFTy outputs}

\item{remove_orphans}{logical. Whether to remove "orphan" paths, i.e. paths
that do not start at time = 0. \code{TRUE} by default.}
}
\value{
object of class \code{"HeFTy"}, i.e. \code{list} of the individual paths,
the constraints, the weighted mean path and the grain summary statistics.
}
\description{
Read time, temperature and GOF data of modeled thermal history paths out of
HeFTy output .txt file
}
\examples{
path2myfile <- system.file("112-73_30_H1_50-inv.txt", package = "thermoclustr")
read_hefty(path2myfile)
}
